fix: structure and models
